playlist/requirements.txt
Pillow
argparse
pyyaml
requests
spotipy